requirements.txt
certifi==2021.5.30
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.4
fastecdsa==2.2.2
filelock==3.7.1
fonttools==4.44.0
idna==3.4
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
platformdirs==2.5.2
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.0.7
virtualenv==20.14.1
zipp==3.17.0
